✅ Separate logout functionality
